Split mission cards modules
