+++ dashboard/src/components/mission-cards.ts
@@ -1,38 +1,14 @@
 import { html } from 'htm/preact'
-import { Card } from './common/card'
-import { ProvenanceStrip } from './common/provenance-strip'
-import { openAgentDetail } from './agent-detail'
-import {
-  missionBriefing,
-  missionBriefingError,
-  missionBriefingLoading,
-  refreshMissionBriefing,
-} from '../mission-store'
-import { workflowActionLabel } from '../workflow-context'
-import type {
-  DashboardMissionAttentionQueueItem,
-  DashboardMissionSessionBrief,
-} from '../types'
 import {
   toneClass,
   relativeTime,
-  statusLabel,
-  missionTargetTypeLabel,
-  signalClassLabel,
-  actionModeLabel,
-  actionTargetLabel,
-  toggleAttention,
-  toggleSession,
-  openIncidentIntervene,
-  openIncidentCommand,
-  openActionIntervene,
-  openActionCommand,
-  attentionAsIncident,
 } from './mission-utils'
 
 // Re-export from split files for consumers importing from './mission-cards'
 export { SessionBriefCard, SessionDetailCard } from './mission-session-cards'
 export { AgentBriefCard, KeeperBriefCard, InternalSignalCard } from './mission-agent-cards'
+export { MissionBriefingCard } from './mission-briefing-card'
+export { AttentionCard } from './mission-attention-card'
 
 export function MissionContextBar({
   cluster,
@@ -87,224 +63,6 @@ export function SummaryStat({
       <span class="mission-stat-label">${label}</span>
       <strong class="mission-stat-value">${value}</strong>
       <small class="mission-stat-detail">${detail}</small>
-    </article>
-  `
-}
-
-export function MissionBriefingCard() {
-  const briefing = missionBriefing.value
-  const briefingTone = toneClass(briefing?.status ?? (missionBriefingError.value ? 'bad' : 'warn'))
-  const showEmpty = !briefing || briefing.sections.length === 0
-  const retryNeedsForce =
-    briefing?.status === 'error'
-    || (briefing?.status === 'unavailable' && !briefing?.cached)
-
-  return html`
-    <${Card} title="판단 레이어" class="mission-briefing-card" semanticId="mission.model_briefing">
-      <div class="mission-section-head">
-        <h3>왜 그렇게 보이나</h3>
-        <p>사회 truth를 읽은 뒤에만 별도 판단 결과를 참고하고, 근거는 접어서 둡니다.</p>
-        <${ProvenanceStrip}
-          items=${[
-            { kind: 'narrative' },
-            { kind: 'fallback', label: 'fallback on failure' },
-          ]}
-        />
-      </div>
-
-      <div class="mission-briefing-meta">
-        <span class="command-chip ${briefingTone}">
-          ${statusLabel(briefing?.status ?? (missionBriefingError.value ? 'error' : 'loading'))}
-        </span>
-        ${briefing?.model ? html`<span class="command-chip">${briefing.model}</span>` : null}
-        ${briefing?.generated_at ? html`<span class="command-chip">${relativeTime(briefing.generated_at)}</span>` : null}
-        ${briefing?.cached ? html`<span class="command-chip">캐시</span>` : null}
-        ${briefing?.stale ? html`<span class="command-chip warn">오래됨</span>` : null}
-        ${briefing?.refreshing ? html`<span class="command-chip warn">갱신 중</span>` : null}
-      </div>
-
-      ${missionBriefingError.value ? html`<div class="empty-state error">${missionBriefingError.value}</div>` : null}
-      ${briefing?.error ? html`<div class="empty-state error">${briefing.error}</div>` : null}
-      ${briefing?.summary ? html`<div class="mission-inline-note">${briefing.summary}</div>` : null}
-      ${briefing?.last_error && !briefing.error
-        ? html`<div class="mission-inline-note">최근 갱신 실패: ${briefing.last_error}</div>`
-        : null}
-
-      ${briefing && briefing.sections.length > 0
-        ? html`
-            <div class="mission-briefing-grid">
-              ${briefing.sections.slice(0, 3).map(section => html`
-                <article class="mission-briefing-section ${toneClass(section.status)}">
-                  <div class="mission-card-head">
-                    <strong>${section.label}</strong>
-                    <div class="mission-briefing-section-chips">
-                      <span class="command-chip ${toneClass(section.status)}">${statusLabel(section.status)}</span>
-                      ${signalClassLabel(section.signal_class)
-                        ? html`<span class="command-chip ${section.signal_class === 'mixed' ? 'warn' : ''}">${signalClassLabel(section.signal_class)}</span>`
-                        : null}
-                      ${section.evidence_quality ? html`<span class="command-chip">${section.evidence_quality}</span>` : null}
-                    </div>
-                  </div>
-                  <p>${section.summary}</p>
-                  ${section.evidence.length > 0
-                    ? html`
-                        <details class="mission-card-disclosure compact">
-                          <summary>근거 보기</summary>
-                          <div class="mission-pill-row">
-                            ${section.evidence.map(item => html`<span class="mission-pill">${item}</span>`)}
-                          </div>
-                        </details>
-                      `
-                    : null}
-                </article>
-              `)}
-            </div>
-          `
-        : (!missionBriefingLoading.value && !missionBriefingError.value && showEmpty
-            ? html`
-                <div class="empty-state">
-                  ${briefing?.status === 'pending'
-                    ? '최신 스냅샷으로 브리핑을 생성 중입니다. 마지막 성공 결과가 생기면 자동으로 다시 읽습니다.'
-                    : '판단 결과가 아직 없습니다.'}
-                </div>
-              `
-            : null)}
-
-      ${briefing && briefing.metadata_gaps.length > 0
-        ? html`
-            <details class="mission-card-disclosure compact mission-briefing-gaps">
-              <summary>관측 공백 (${briefing.metadata_gap_count ?? briefing.metadata_gaps.length})</summary>
-              <div class="mission-list-stack">
-                ${briefing.metadata_gaps.map(item => html`
-                  <article class="mission-briefing-gap ${item.severity === 'watch' ? 'warn' : ''}">
-                    <div class="mission-card-head">
-                      <strong>${missionTargetTypeLabel(item.scope_type)}${item.scope_id ? ` · ${item.scope_id}` : ''}</strong>
-                      <span class="command-chip ${item.severity === 'watch' ? 'warn' : ''}">${statusLabel(item.severity)}</span>
-                    </div>
-                    <p>${item.summary}</p>
-                  </article>
-                `)}
-              </div>
-            </details>
-          `
-        : null}
-
-      <div class="mission-card-actions">
-        <button class="control-btn ghost" onClick=${() => { void refreshMissionBriefing(retryNeedsForce) }} disabled=${missionBriefingLoading.value}>
-          ${missionBriefingLoading.value ? '응답 기다리는 중…' : '판단 다시 읽기'}
-        </button>
-        <button class="control-btn ghost" onClick=${() => { void refreshMissionBriefing(true) }} disabled=${missionBriefingLoading.value}>
-          강제 갱신
-        </button>
-      </div>
-    <//>
-  `
-}
-
-export function AttentionCard({
-  item,
-  selected,
-  sessionLookup,
-}: {
-  item: DashboardMissionAttentionQueueItem
-  selected: boolean
-  sessionLookup: Map<string, DashboardMissionSessionBrief>
-}) {
-  const incident = attentionAsIncident(item)
-  const linkedSessions = item.related_session_ids
-    .map(id => sessionLookup.get(id))
-    .filter((row): row is DashboardMissionSessionBrief => row != null)
-  const action = item.top_action ?? null
-
-  return html`
-    <article class="mission-attention-card ${toneClass(action?.severity ?? item.severity)} ${selected ? 'is-selected' : ''}">
-      <button class="mission-card-select" onClick=${() => toggleAttention(item.id)}>
-        <div class="mission-card-head">
-          <div>
-            <strong>${item.summary}</strong>
-            <div class="mission-card-target">${missionTargetTypeLabel(item.target_type)}${item.target_id ? ` · ${item.target_id}` : ''}</div>
-          </div>
-          <span class="command-chip ${toneClass(action?.severity ?? item.severity)}">${action ? actionModeLabel(action) : item.severity}</span>
-        </div>
-
-        <div class="mission-fact-grid">
-          <div class="mission-fact-tile">
-            <span>영향 세션</span>
-            <strong>${item.related_session_ids.length}</strong>
-            <small>${item.related_session_ids.slice(0, 2).join(', ') || '없음'}</small>
-          </div>
-          <div class="mission-fact-tile">
-            <span>영향 에이전트</span>
-            <strong>${item.related_agent_names.length}</strong>
-            <small>${item.related_agent_names.slice(0, 3).join(', ') || '없음'}</small>
-          </div>
-          <div class="mission-fact-tile">
-            <span>최근 신호</span>
-            <strong>${item.last_seen_at ? relativeTime(item.last_seen_at) : '기록 없음'}</strong>
-            <small>${missionTargetTypeLabel(item.target_type)}</small>
-          </div>
-          <div class="mission-fact-tile">
-            <span>다음 액션</span>
-            <strong>${action ? workflowActionLabel(action.action_type) : '판단 필요'}</strong>
-            <small>${action ? actionTargetLabel(action) : '추천 액션 없음'}</small>
-          </div>
-        </div>
-      </button>
-
-      ${action ? html`<div class="mission-inline-note">${action.reason}</div>` : null}
-
-      <details class="mission-card-disclosure">
-        <summary>연결된 흐름 보기</summary>
-        ${linkedSessions.length > 0
-          ? html`
-              <div class="mission-link-list">
-                ${linkedSessions.slice(0, 4).map(session => html`
-                  <button class="mission-link-row" onClick=${() => toggleSession(session.session_id)}>
-                    <strong>${session.goal}</strong>
-                    <span>${statusLabel(session.status)} · ${session.last_event_summary ?? '최근 사건 없음'}</span>
-                  </button>
-                `)}
-              </div>
-            `
-          : html`<div class="empty-state">직접 연결된 세션이 아직 없습니다.</div>`}
-
-        ${item.related_agent_names.length > 0
-          ? html`
-              <div class="mission-pill-row">
-                ${item.related_agent_names.slice(0, 8).map(name => html`
-                  <button class="mission-pill action" onClick=${() => openAgentDetail(name)}>${name}</button>
-                `)}
-              </div>
-            `
-          : null}
-
-        ${item.evidence_preview.length > 0
-          ? html`
-              <details class="mission-card-disclosure compact">
-                <summary>근거 미리보기</summary>
-                <div class="mission-evidence-list">
-                  ${item.evidence_preview.map(text => html`<span>${text}</span>`)}
-                </div>
-              </details>
-            `
-          : null}
-      </details>
-
-      <div class="mission-card-actions">
-        ${action
-          ? html`
-              <button class="control-btn ghost" onClick=${() => openActionIntervene(action, incident, '상황판 주의 신호')}>
-                이 액션으로 개입 열기
-              </button>
-              <button class="control-btn ghost" onClick=${() => openActionCommand(action, incident, '상황판 주의 신호')}>
-                원인 보기
-              </button>
-            `
-          : html`
-              <button class="control-btn ghost" onClick=${() => openIncidentIntervene(incident)}>이 이슈로 개입 열기</button>
-              <button class="control-btn ghost" onClick=${() => openIncidentCommand(incident)}>이 이슈의 원인 보기</button>
-            `}
-      </div>
     </article>
   `
 }
